# Optimize injection and operator jobs