--- a/istioctl/pkg/verifier/verifier.go
+++ b/istioctl/pkg/verifier/verifier.go
@@ -215,7 +215,7 @@ func (v *StatusVerifier) verifyInstall() error {
 func (v *StatusVerifier) verifyPostInstallIstioOperator(iop *v1alpha1.IstioOperator, filename string) (int, int, error) {
 	t := translate.NewTranslator()
 
-	cp, err := controlplane.NewIstioControlPlane(iop.Spec, t)
+	cp, err := controlplane.NewIstioControlPlane(iop.Spec, t, nil)
 	if err != nil {
 		return 0, 0, err
 	}