--- a/operator/pkg/util/yaml.go
+++ b/operator/pkg/util/yaml.go
@@ -86,6 +86,17 @@ func UnmarshalWithJSONPB(y string, out proto.Message, allowUnknownField bool) er
 
 // OverlayTrees performs a sequential JSON strategic of overlays over base.
 func OverlayTrees(base map[string]interface{}, overlays ...map[string]interface{}) (map[string]interface{}, error) {
+	needsOverlay := false
+	for _, o := range overlays {
+		if len(o) > 0 {
+			needsOverlay = true
+			break
+		}
+	}
+	if !needsOverlay {
+		// Avoid expensive overlay if possible
+		return base, nil
+	}
 	bby, err := yaml.Marshal(base)
 	if err != nil {
 		return nil, err