… manifests/charts/default/templates/mutatingwebhook.yaml
… manifests/charts/istio-control/istio-discovery/templates/mutatingwebhook.yaml
… manifests/charts/istiod-remote/templates/mutatingwebhook.yaml
… operator/cmd/mesh/manifest_shared_test.go
… operator/cmd/mesh/testdata/manifest-generate/data-snapshot.tar.gz
… operator/cmd/mesh/testdata/manifest-generate/output/all_on.golden-show-in-gh-pull-request.yaml
… operator/cmd/mesh/testdata/manifest-generate/output/flag_output.golden.yaml
… operator/cmd/mesh/testdata/manifest-generate/output/flag_output_set_values.golden.yaml
… operator/cmd/mesh/testdata/manifest-generate/output/helm_values_enablement.golden.yaml
… operator/cmd/mesh/testdata/manifest-generate/output/ingressgateway_k8s_settings.golden.yaml
… operator/cmd/mesh/testdata/manifest-generate/output/install_package_path.golden.yaml
… operator/cmd/mesh/testdata/manifest-generate/output/pilot_default.golden.yaml
… operator/cmd/mesh/testdata/manifest-generate/output/pilot_k8s_settings.golden.yaml
… operator/cmd/mesh/testdata/manifest-generate/output/pilot_override_values.golden.yaml
… operator/cmd/mesh/testdata/manifest-generate/output/sidecar_template.golden.yaml
… operator/cmd/mesh/testdata/operator/output/operator-dump.json
… operator/cmd/mesh/testdata/operator/output/operator-dump.yaml
… operator/cmd/mesh/testdata/operator/output/operator-init.yaml
… operator/cmd/mesh/testdata/profile-dump/output/all_off.txt
… operator/cmd/mesh/testdata/profile-dump/output/all_off.yaml
… operator/pkg/controlplane/control_plane_test.go
… operator/pkg/util/yaml_test.go
… pkg/kube/inject/inject_test.go
… pkg/kube/inject/testdata/inputs/custom-template.yaml.34.mesh.gen.yaml
… pkg/kube/inject/testdata/inputs/custom-template.yaml.34.template.gen.yaml
… pkg/kube/inject/testdata/inputs/custom-template.yaml.34.values.gen.yaml
… pkg/kube/inject/testdata/inputs/default.mesh.gen.yaml
… pkg/kube/inject/testdata/inputs/default.template.gen.yaml
… pkg/kube/inject/testdata/inputs/default.values.gen.yaml
… pkg/kube/inject/testdata/inputs/enable-core-dump.yaml.5.mesh.gen.yaml