For https://github.com/istio/istio/issues/37552

This may seem like premature optimizations, but these jobs are **incredibly** slow and use 100% cpu the entire time. This, in turn, makes other tests slower/flaky.

In real world, we also see injector using 5% memory total

Changes done here:
* In injector, precompute values map/struct and pre-parse templates.
* In webhook yaml, avoid deep copy of the entire values.yaml tree
* Add ability to pre-filter manifests in operator generation tests. This allows skipping a bunch of resources we don't care about in tests
* Add ability to skip Prune, which is expensive
* In injector tests, store the rendered manifests in files to avoid recomputing each time


This PR should have no user-facing changes

Before:
ok         istio.io/istio/operator/cmd/mesh        163.397s
ok        istio.io/istio/pkg/kube/inject  40.178s

After:
ok         istio.io/istio/operator/cmd/mesh        90.035s
ok        istio.io/istio/pkg/kube/inject  11.369s

Note: these jobs now *temporarily* run on 15 CPU dedicated nodes (up from 5 core shared). When running on even smaller nodes, such as 3 nodes, the before runtime was much higher - I saw 40 **minutes** in my tests