@howardjohn 2022-03-01
Todo: precompute manifest for injection. Also can/should use RenderManifestFiltered?
Look into why operator is so slow