@istio-testing 2022-03-01
@howardjohn: The following tests **failed**, say `/retest` to rerun all failed tests or `/retest-required` to rerun all mandatory failed tests:

Test name | Commit | Details | Required | Rerun command
--- | --- | --- | --- | ---
integ-ipv6_istio | 12ffbda2b4e92113eb96c7f08c9b5107f06d40c0 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/37607/integ-ipv6_istio/1498766943300096000) | true | `/test integ-ipv6_istio`
integ-helm_istio | 12ffbda2b4e92113eb96c7f08c9b5107f06d40c0 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/37607/integ-helm_istio/1498766948865937408) | true | `/test integ-helm_istio`
integ-operator-controller_istio | 12ffbda2b4e92113eb96c7f08c9b5107f06d40c0 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/37607/integ-operator-controller_istio/1498766943333650432) | true | `/test integ-operator-controller_istio`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->